Separate drawable pieces of each HitObject type
