+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/DrawableSlideTap.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Objects.Types;
-using osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces;
+using osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides;
 
 namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 {
