+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/Pieces/Slides/StarPiece.cs
@@ -5,7 +5,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osuTK;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
+namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides
 {
     public class StarPiece : CompositeDrawable
     {
